Add initial eligibility flow
The idea behind the eligibility flow implementation is that it should be form-based, to minimize the amount of JavaScript theoretically possible. I am not using Next.js's experimental server components, but I would like to investigate this. For now, the only flow which requires JavaScript is the building type selection mechanism, since the design says that follow-up questions may be asked.

No styling is yet included. This is as bare-bones as it gets.